# openapi/firewall.v1.yaml
openapi: 3.1.0
info:
  title: SIGMACODE Neural Firewall API
  version: 1.0.0
  description: >
    Core-Endpoints für AI-Validation (Guardrail-Generator), Text-/File-Scanning mit Redaction
    und Policy-Sync (Edge/On-Prem). Integriert die Firewall-Entscheidungen (allow/block/sanitize),
    Shadow-Mode-Unterstützung und Request-Korrelation (x-request-id).
servers:
  - url: https://api.sigmacode.ai
    description: Production
  - url: https://staging.api.sigmacode.ai
    description: Staging
  - url: http://localhost:3000
    description: Local Dev
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XRequestId:
      name: x-request-id
      in: header
      required: false
      description: Client- oder Server-generierte Korrelations-ID
      schema:
        type: string
...
  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        code: { type: string }
        details:
          type: object
          additionalProperties: true
      required: [error, code]
...
    DecisionMeta:
      type: object
      properties:
        requestId: { type: string }
        backend: { type: string, description: 'e.g., superagent|dify|edge' }
        mode: { type: string, enum: [enforce, shadow, off] }
        latencyMs: { type: integer }
        policy: { type: string }
        reason: { type: string }
        threats:
          type: array
          items:
            type: string
      required: [mode]
...
    Guardrail:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        type:
          type: string
          enum: [input_filter, output_filter, context_check, format_validation]
        condition: { type: string }
        action:
          type: string
          enum: [block, sanitize, warn, transform]
        priority: { type: integer, default: 1 }
        enabled: { type: boolean, default: true }
        metadata:
          type: object
          additionalProperties: true
...
    Policy:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        version: { type: string }
        mode:
          type: string
          enum: [enforce, shadow, off]
        rules:
          type: array
          items: { $ref: '#/components/schemas/Guardrail' }
        updatedAt: { type: string, format: date-time }
      required: [id, name, version, mode, rules]
    PolicyCreate:
      type: object
      properties:
        name: { type: string }
        mode: { type: string, enum: [enforce, shadow, off], default: enforce }
        rules:
          type: array
          items: { $ref: '#/components/schemas/Guardrail' }
      required: [name, mode, rules]
...
    PolicyResponse:
      type: object
      properties:
        policy: { $ref: '#/components/schemas/Policy' }
    PoliciesListResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Policy' }
        total: { type: integer }
      required: [data, total]
    FirewallConfig:
      type: object
      properties:
        isEnabled: { type: boolean }
        mode: { type: string, enum: [enforce, shadow, off] }
        defaultPolicyId: { type: string, nullable: true }
        sampling:
          type: object
          properties:
            shadowSampleRate:
              type: number
              description: '0..1 Anteil, der im Shadow-Mode gespiegelt wird'
        updatedAt: { type: string, format: date-time }
      required: [isEnabled, mode]
    ConfigResponse:
      type: object
      properties:
        config: { $ref: '#/components/schemas/FirewallConfig' }
        meta: { $ref: '#/components/schemas/DecisionMeta' }
      required: [config]
    FirewallConfigUpdate:
      type: object
      properties:
        isEnabled: { type: boolean }
        mode: { type: string, enum: [enforce, shadow, off] }
        defaultPolicyId: { type: string, nullable: true }
        sampling:
          type: object
          properties:
            shadowSampleRate:
              type: number
...
paths:
...
  /firewall/policies:
    get:
      tags: [policies]
      summary: List policies
      operationId: listPolicies
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 100, minimum: 1, maximum: 500 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        '200':
          description: Policies list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PoliciesListResponse' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    post:
      tags: [policies]
      summary: Create policy
      operationId: createPolicy
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/XRequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PolicyCreate' }
      responses:
        '201':
          description: Policy created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PolicyResponse' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /firewall/config:
    get:
      tags: [policies]
      summary: Get firewall config
      operationId: getFirewallConfig
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/XRequestId'
      responses:
        '200':
          description: Config
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConfigResponse' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    put:
      tags: [policies]
      summary: Update firewall config
      operationId: updateFirewallConfig
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/XRequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FirewallConfigUpdate' }
      responses:
        '200':
          description: Updated config
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConfigResponse' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...